+++ packages/editor/editor-wikimarkup-transformer/src/__tests__/unit/parser/other.ts
@@ -1,0 +1,25 @@
+import WikiMarkupTransformer from '../../../index';
+
+describe('JIRA wiki markup - Other', () => {
+  const testCases: Array<[string, string]> = [
+    [
+      'should begin with hardBreak if it begins with a new line',
+      `
+foo`,
+    ],
+    [
+      'should process second empty line as a new paragraph',
+      `foo
+
+
+bar`,
+    ],
+  ];
+
+  for (const [testCaseDescription, markup] of testCases) {
+    it(testCaseDescription, () => {
+      const transformer = new WikiMarkupTransformer();
+      expect(transformer.parse(markup)).toMatchSnapshot();
+    });
+  }
+});