+++ packages/media/media-picker/src/service/uploadServiceFactory.ts
@@ -9,7 +9,6 @@ import {
   UploadParams,
 } from '..';
 import { NewUploadServiceImpl } from './newUploadServiceImpl';
-import { OldUploadServiceImpl } from './uploadService';
 
 export type UploadServiceEventPayloadTypes = {
   readonly 'files-added': UploadsStartEventPayload;
@@ -46,16 +45,11 @@ export class UploadServiceFactory {
     context: Context,
     tenantUploadParams: UploadParams,
     uploadParams?: UploadParams,
-    useNewUploadService: boolean = false,
   ): UploadService {
-    if (useNewUploadService) {
-      return new NewUploadServiceImpl(
-        context,
-        tenantUploadParams,
-        uploadParams,
-      );
-    } else {
-      return new OldUploadServiceImpl(context, uploadParams);
-    }
+    return new NewUploadServiceImpl(
+      context,
+      tenantUploadParams,
+      uploadParams,
+    );
   }
 }