+++ packages/core/tooltip/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/tooltip
 
+## 12.0.11
+- [patch] tooltip will not appear when content is undefined null or an empty string [239b448](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/239b448)
+
 ## 12.0.9
 - [patch] Updated dependencies [df22ad8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/df22ad8)
   - @atlaskit/theme@6.0.0