+++ packages/editor/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/editor-test-helpers
 
+## 6.0.4
+- [patch] Check current selected nodes before change node selection when interacting with extensions. ED-5199 [bb15908](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bb15908)
+
 ## 6.0.3
 - [patch] Updated dependencies [b12f7e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b12f7e6)
   - @atlaskit/editor-common@14.0.11